Organise statistics files and wire up the boolean statistics
