+++ CumulusMX/CumulusMX.csproj
@@ -27,9 +27,4 @@
     </None>
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Data\Statistics\Double\" />
-    <Folder Include="Data\Statistics\Unit\" />
-  </ItemGroup>
-
 </Project>
