+++ CumulusMX/Data/Statistics/Double/RollingStatisticDouble.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace CumulusMX.Data
+namespace CumulusMX.Data.Statistics.Double
 {
     public class RollingStatisticDouble
     {
